requirements.txt
opencv-python
ultralytics
pyodbc
numpy
python-dotenv